requirements.txt
streamlit
Pillow
exif
folium
streamlit-folium
